fix: update shopping page ui
