+++ src/components/tasks/ui/modal/TaskStatsModal.tsx
@@ -70,6 +70,9 @@ const TaskStatsModal: React.FC<TaskStatsModalProps> = ({
                   <X size={20} />
                 </button>
               </div>
+              <h3 className="text-lg font-semibold text-gray-900 mb-4">
+                Quick Stats
+              </h3>
               <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
                 {stats.map((stat, index) => (
                   <motion.div
